+++ pyfilter/inference/base.py
@@ -5,7 +5,7 @@ import torch
 from ..filters import BaseFilter
 from .logging import DefaultLogger
 from .state import AlgorithmState
-from .context import ParameterContext
+from .context import InferenceContext
 
 
 class BaseAlgorithm(ABC):
@@ -13,7 +13,7 @@ class BaseAlgorithm(ABC):
     Abstract base class for algorithms.
     """
 
-    def __init__(self, filter_: BaseFilter, context: ParameterContext = None):
+    def __init__(self, filter_: BaseFilter, context: InferenceContext = None):
         """
         Initializes the :class:`BaseFilterAlgorithm` class.
 
@@ -25,7 +25,7 @@ class BaseAlgorithm(ABC):
         super().__init__()
 
         self._filter = filter_
-        self.context: ParameterContext = context or ParameterContext.get_context()
+        self.context: InferenceContext = context or InferenceContext.get_context()
 
     @property
     def filter(self) -> BaseFilter: