+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # v0.25.1
  - Reworks `ParameterContext` logic by removing the requirements for the context being on the stack in order to register parameters.
  - Adds fixed-lag smoothing.
+ - Renames `ParameterContext` to `InferenceContext`
 
 # v0.24.11
  - Adds support for using QMC points in `SMC2`